requirements.txt
fastapi==0.110.0
httpx==0.27.0
pandas==2.2.1
pydantic==2.6.4
pydantic_settings==2.2.1
pytest==8.1.1
PyYAML==6.0.1
uvicorn==0.29.0
